renamed constants_and_variables.h
