+++ constants.h
